cart checks if products have sold out
