+++ components/Cart.js
@@ -1,30 +1,13 @@
-import { useState } from 'react';
-import Link from 'next/link';
 import { DialogOverlay, DialogContent } from '@reach/dialog';
 import styled, { keyframes } from 'styled-components';
 
 import Icon from './Icon';
 import UnstyledButton from './UnstyledButton';
 import VisuallyHidden from './VisuallyHidden';
-import { useCart } from '../hooks/useCart';
-import formatMoney from '../util/formatMoney';
-import CartItem from './CartItem';
-import stripeCheckout from '../util/stripeCheckout';
 import { QUERIES } from '../util/constants';
+import CartBody from './CartBody';
 
 function Cart({ isOpen, onDismiss }) {
-    const [working, setWorking] = useState(false);
-    const { cart } = useCart();
-    let total = 0;
-    const slugs = Object.keys(cart);
-
-    const handleClick = async (e) => {
-        e.preventDefault();
-        setWorking(true);
-        await stripeCheckout(cart);
-        setWorking(false);
-    };
-
     return (
         <Overlay isOpen={isOpen} onDismiss={onDismiss}>
             <Backdrop />
@@ -36,35 +19,7 @@ function Cart({ isOpen, onDismiss }) {
                     </CloseButton>
                     <Title>Your Cart</Title>
                 </Header>
-                <CartBody>
-                    <Items>
-                        {slugs.map((key) => {
-                            const { slug, price } = cart[key];
-                            total += price;
-                            return <CartItem key={slug} slug={slug} />;
-                        })}
-                    </Items>
-                    {total > 0 ? (
-                        <Checkout>
-                            <div>
-                                Subtotal ({slugs.length} item{slugs.length > 1 && 's'}):{' '}
-                                <Money>{formatMoney(total)}</Money>
-                            </div>
-                            <Buttons>
-                                <BuyButton type='button' onClick={handleClick} disabled={working}>
-                                    Check Out
-                                </BuyButton>
-                                <Link href='/'>
-                                    <Button type='button' onClick={onDismiss}>
-                                        Continue Shopping
-                                    </Button>
-                                </Link>
-                            </Buttons>
-                        </Checkout>
-                    ) : (
-                        <div>Your Shopping Cart is empty</div>
-                    )}
-                </CartBody>
+                <CartBody onDismiss={onDismiss} />
             </Content>
         </Overlay>
     );
@@ -114,7 +69,7 @@ const Content = styled(DialogContent)`
     background: white;
     width: 425px;
     height: 100%;
-    overflow: scroll;
+    overflow-y: scroll;
 
     @media (prefers-reduced-motion: no-preference) {
         animation: ${slideIn} 500ms both cubic-bezier(0, 0.6, 0.32, 1.06);
@@ -140,60 +95,6 @@ const CloseButton = styled(UnstyledButton)`
 `;
 const Title = styled.h2`
     font-size: 1.5rem;
-`;
-
-const CartBody = styled.div`
-    display: flex;
-    flex-direction: column;
-    padding: 32px;
-    gap: 32px;
-`;
-
-const Items = styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-`;
-
-const Checkout = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 8px;
-`;
-
-const Button = styled.button`
-    padding: 8px 24px;
-    font-size: 1rem;
-    font-weight: var(--bold);
-    color: white;
-    background: var(--gray-700);
-    border-radius: 8px;
-    border: none;
-    width: 120px;
-
-    &:hover,
-    &:focus {
-        background: var(--gray-900);
-    }
-`;
-
-const BuyButton = styled(Button)`
-    background: hsl(0, 50%, 50%);
-
-    &:hover,
-    &:focus {
-        background: hsl(0, 80%, 50%);
-    }
-`;
-
-const Buttons = styled.div`
-    display: flex;
-    gap: 16px;
-`;
-
-const Money = styled.span`
-    font-weight: var(--bold);
 `;
 
 export default Cart;
